requirements.txt
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
python-dotenv
pandas
